+++ firebase/functions/package.json
@@ -27,8 +27,7 @@
     "@google-cloud/billing": "3.1.3",
     "@google-cloud/pubsub": "3.2.1",
     "firebase-admin": "11.3.0",
-    "firebase-functions": "4.2.0",
-    "jest": "28.0.0"
+    "firebase-functions": "4.2.0"
   },
   "devDependencies": {
     "@types/node": "16.18.3",
@@ -37,7 +36,6 @@
     "eslint": "8.28.0",
     "eslint-config-google": "0.14.0",
     "eslint-plugin-import": "2.26.0",
-    "firebase-functions-test": "3.0.0",
     "typescript": "4.9.3",
     "yarn-upgrade-all": "0.7.1"
   },
