+++ src/pages/profile/edit.tsx
@@ -1,10 +1,9 @@
 import { ReactElement } from 'react'
-import { useRouter } from 'next/router'
 
 import ContactForm from '@/components/Contact/contactform'
+import NavButtons from '@/components/Contact/navbuttons'
 import Layout from '@/components/Layout'
 import { withProtected } from '@/components/PrivateRoute'
-import Button from '@/components/UI/button'
 import Spinner from '@/components/UI/loadingSpinner'
 import useUser, { useMutateUser } from '@/hooks/user'
 import { NextPageWithLayout } from '@/pages/_app'
@@ -12,7 +11,6 @@ import { NextPageWithLayout } from '@/pages/_app'
 const Profile: NextPageWithLayout = () => {
   const { data, isLoading } = useUser()
   const { mutate: mutateUser } = useMutateUser()
-  const router = useRouter()
 
   if (isLoading)
     return (
@@ -27,17 +25,7 @@ const Profile: NextPageWithLayout = () => {
 
   return (
     <div className='main-style'>
-      <div className='my-4 ml-6 h-14 w-max text-center'>
-        <Button
-          type='button'
-          size='medium'
-          onClick={() => {
-            router.back()
-          }}
-        >
-          Back
-        </Button>
-      </div>
+      <NavButtons />
       <ContactForm contact={user} mutate={mutateUser} />
     </div>
   )
